requirements.txt
streamdeck
python-socketio[client]
pillow
pyinstaller